Dev 12/3 (#3)
* added async loading of the KMLLayers to decrease load time, added alert dialogue to recyclerList on long click for deleting issue.
* lots of little clean up items

+++ app/src/main/java/com/skoorc/atvolunteeraid/view/ListFragment.kt
@@ -1,7 +1,9 @@
 package com.skoorc.atvolunteeraid.view
 
+import android.app.AlertDialog
 import android.content.Intent
 import android.os.Bundle
+import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
@@ -29,6 +31,7 @@ class ListFragment: Fragment(), LocationListAdapter.OnItemClickListener, Locatio
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
+        activity?.setTitle(R.string.title_activity_report_list)
         val view = inflater.inflate(R.layout.fragment_list_view, container, false)
         val recyclerView = view.findViewById<RecyclerView>(R.id.recyclerview)
         adapter = LocationListAdapter(view.context, this)
@@ -57,7 +60,7 @@ class ListFragment: Fragment(), LocationListAdapter.OnItemClickListener, Locatio
     }
 
     override fun onItemClick(recyclerPosition: Int, id: Int, location: String) {
-        //TODO Navigation to the map focused on the targeted pin. Look for a more refined way of
+        //TODO Look for a more refined way of
         // doing this than passing extras in through the intent
         Toast.makeText(context, "Let's go to the MAP!", Toast.LENGTH_SHORT).show()
         val intent = Intent(requireContext(), MapsActivity::class.java)
@@ -69,7 +72,19 @@ class ListFragment: Fragment(), LocationListAdapter.OnItemClickListener, Locatio
     override fun onItemLongClick(recyclerPosition: Int, idValue: Int) {
         //TODO Make this pop up a verification that the item has been fixed/removed/Cleaned up
         //TODO Schema update should include resolved true/false. remove deletion (keep all entries for analytics eventually)
-        Toast.makeText(context, "item LOOOONGGGGG clicked $recyclerPosition, ID: $idValue", Toast.LENGTH_SHORT).show()
-        locationViewModel.deleteById(idValue)
+        Log.d(TAG,"Recycler item long clicked, position: $recyclerPosition, ID: $idValue")
+        val dialogueBuilder = AlertDialog.Builder(context)
+        dialogueBuilder.setMessage("Has the issue been cleaned up or fixed??")
+            .setCancelable(false)
+            .setPositiveButton("Yup, Issue resolved") { dialogue, _ ->
+                locationViewModel.deleteById(idValue)
+                dialogue.dismiss()
+            }
+            .setNegativeButton("Cancel") { dialogue, _ ->
+                dialogue.cancel()
+            }
+        val alert = dialogueBuilder.create()
+        alert.setTitle("Issue resolution")
+        alert.show()
     }
 }
